requirements.txt
dnspython
schedule
requests
python-dotenv